Fix dependency conflicts and compilation errors in rustyjack-ui

+++ rustyjack-ui/src/display.rs
@@ -84,7 +84,7 @@ impl Display {
         backlight.set_value(1)?;
 
         let mut delay = Delay {};
-        let mut lcd = ST7735::new(spi, dc, rst, true, false, LCD_WIDTH, LCD_HEIGHT);
+        let mut lcd = ST7735::new(spi, dc, rst, true, false, LCD_WIDTH as u32, LCD_HEIGHT as u32);
         lcd.init(&mut delay)?;
         lcd.set_orientation(&Orientation::Portrait)?;
         lcd.set_offset(LCD_OFFSET_X, LCD_OFFSET_Y);
